[build: update (un)install]
Remove the all target, which was used for building a .vim directory,
replacing it with the PREFIX variable to control the install location.

Fix cyclic links, which occurred if install is called multiple times,
locating the build directory within the repository. Since the build
directory was not excluded, it was considered as a plugin directory,
creating links from files within the build directory to themselves.
This is fixed, by excluding the build directory from the plugin
directory list.

Fix the creation of untracked files within the plugins, which occurred
if install is called multiple times. Since the first call created
symlinks the second call to install will allocate the new symlinks below
the ones from the first call. Since the first symlinks already point to
the plugin directories, this creates files within the plugins.
To fix it, existing symlinks are removed before allocating new ones.